Migrated to WifiSensorManager

+++ app/src/main/java/uk/ac/kent/eda/jb956/sensorlibrary/data/WifiData.java
@@ -6,7 +6,7 @@ package uk.ac.kent.eda.jb956.sensorlibrary.data;
  */
 public class WifiData extends SensorData {
     public String bssid;
-    public int rssi;
+    public double rssi;
     public double distanceEstimate;
 
 }
